delete test helpers moved to shared/test_helpers
