+++ src/client/components/draft/shared/test_helpers.ts
